when using itertuples and loc to modify a cell value
... be sure to use row.Index  and not row.index !!! else no change is
recorded ! Also "set_value" looks better.

Plotting OWDs is quite timeconsuming so we should add an option to plot
them from a cached file (or even cache this automatically !!).
We should also save the score to have an idea how close we are (we might
want to set a threshold) and if several packets are mapped to the same
packet, then map it to the one with highest score only !!