Multiple changes !
Replace --regen by --no-cache. Created a class specific for caching
(cleaner) that I hope to extend to provide intermediate caching (for
instance when plotting OWDs, an intermediate dataframe is
generated with computed OWDs/scores. It would be nice to provide caching
for that.

Also added a tox.ini read by flake8 or pycodestyle (the new name of
    pep8). It allows for longer lines than the default of 80.

The configuration file now reads the [mptcpanalyzer] section instead of
[DEFAULT]